requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ML Dependencies
insightface==0.7.3
onnxruntime-gpu==1.16.3  # ✅ GPU version (auto-falls back to CPU if no CUDA)
opencv-python-headless==4.8.1.78
numpy==1.24.3

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0










# # FastAPI Core
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# python-multipart==0.0.6

# # ML Dependencies
# insightface==0.7.3
# onnxruntime==1.16.3
# opencv-python-headless==4.8.1.78
# numpy==1.24.3

# # Utilities
# pydantic==2.5.0
# python-dotenv==1.0.0
